+++ deployment/cdk/opensearch-service-migration/test/migration-console-stack.test.ts
@@ -1,17 +1,9 @@
 import {createStackComposer} from "./test-utils";
 import {Capture, Match, Template} from "aws-cdk-lib/assertions";
 import {MigrationConsoleStack} from "../lib/service-stacks/migration-console-stack";
-import {ContainerImage} from "aws-cdk-lib/aws-ecs";
-import * as yaml from 'yaml';
-
-class MockDockerImageAsset {
-    imageUri: string;
-    constructor() {
-        this.imageUri = 'mock-image-uri';
-    }
-}
 
 test('Test that IAM policy contains fetch migration IAM statements when fetch migration is enabled', () => {
+    jest.clearAllMocks
     const contextOptions = {
         vpcEnabled: true,
         migrationAssistanceEnabled: true,
@@ -62,55 +54,4 @@ test('Test that IAM policy does not contain fetch migration IAM statements when 
     const iamPassRoleStatement = allStatements.find(statement => statement.Action == "iam:PassRole")
     expect(runTaskStatement).toBeFalsy()
     expect(iamPassRoleStatement).toBeFalsy()
-})
-
-
-test('Test that services yaml parameter is created', () => {
-    // Mock DockerImageAsset to use the mock class
-    jest.mock('aws-cdk-lib/aws-ecr-assets', () => {
-        const originalModule = jest.requireActual('aws-cdk-lib/aws-ecr-assets');
-        return {
-            ...originalModule,
-            DockerImageAsset: MockDockerImageAsset
-        };
-    });
-
-    // Mock ContainerImage.fromDockerImageAsset to return a registry image
-    jest.spyOn(ContainerImage, 'fromDockerImageAsset').mockImplementationOnce(() => ContainerImage.fromRegistry('ServiceImage'));
-
-    const contextOptions = {
-        vpcEnabled: true,
-        migrationAssistanceEnabled: true,
-        migrationConsoleServiceEnabled: true,
-        sourceClusterEndpoint: "https://test-cluster",
-        reindexFromSnapshotServiceEnabled: true,
-        trafficReplayerServiceEnabled: true
-    }
-
-    const stacks = createStackComposer(contextOptions)
-
-    const migrationConsoleStack: MigrationConsoleStack = (stacks.stacks.filter((s) => s instanceof MigrationConsoleStack)[0]) as MigrationConsoleStack
-    const migrationConsoleStackTemplate = Template.fromStack(migrationConsoleStack)
-
-    const valueCapture = new Capture();
-    migrationConsoleStackTemplate.hasResourceProperties("AWS::SSM::Parameter", {
-        Type: "String",
-        Name: Match.stringLikeRegexp("/migration/.*/.*/servicesYamlFile"),
-        Value: valueCapture,
-    });
-    const value = valueCapture.asObject()
-    expect(value).toBeDefined();
-    expect(value['Fn::Join']).toBeInstanceOf(Array);
-    expect(value['Fn::Join'][1]).toBeInstanceOf(Array)
-    // join the strings together to get the yaml file contents
-    const yamlFileContents = value['Fn::Join'][1].join('')
-    expect(yamlFileContents).toContain('source_cluster')
-    expect(yamlFileContents).toContain('target_cluster')
-    expect(yamlFileContents).toContain('metrics_source:\n  cloudwatch:')
-    // Validates that the file can be parsed as valid yaml and has the expected fields
-    const parsedFromYaml = yaml.parse(yamlFileContents);
-    // Validates that the file has the expected fields
-    const expectedFields = ['source_cluster', 'target_cluster', 'metrics_source', 'backfill', 'snapshot', 'metadata_migration', 'replay'];
-    expect(Object.keys(parsedFromYaml).length).toEqual(expectedFields.length)
-    expect(new Set(Object.keys(parsedFromYaml))).toEqual(new Set(expectedFields))
 })
